Updated Auto Deleter

+++ event/auto/autoDelete.js
@@ -1,35 +1,37 @@
 const { EmbedBuilder } = require("discord.js");
 const db = require("../../data/guild/guild");
-const logger = require("../../util/logger.js");
 
 module.exports = {
   name: "messageCreate",
   async execute(message, client) {
-    if (!message.author.bot) return;
-
     try {
-      // Fetch guild data
-      const guildData = await db.findOne({ guildId: message.guild.id });
-      if (!guildData || !guildData.autoDelete.enabled) return;
-
-      // Check if the message is in the configured auto-delete channel
+      let guildData = await db.findOne({ guildId: message.guild.id });
+      if (!guildData || !guildData.autoDelete || !guildData.autoDelete.enabled)
+        return;
       if (message.channel.id !== guildData.autoDelete.id) return;
-
-      // Prevent bot from deleting its own messages
+      if (!guildData || !guildData.autoDelete.enabled) return;
       if (message.author.id === client.user.id) return;
+      if (!message.author.bot) return;
 
-      // Delete the bot message
       await message.delete();
 
-      // Create an embed notification
-      const embed = new EmbedBuilder()
-        .setColor("#FFFFFF")
-        .setTitle("Fabric's Auto Message Deleter")
-        .setDescription(`Deleted message from <@${message.author.id}>`);
+      let notificationMessage =
+        guildData.autoDelete.text || "Default notification message.";
+      if (notificationMessage.includes("CHANNEL")) {
+        notificationMessage = notificationMessage.replace(
+          "CHANNEL",
+          `<#${message.channel.id}>`
+        );
+      }
 
-      await message.channel.send
+      const embed = client.buildEmbed(client, {
+        title: "Auto Deleter",
+        description: notificationMessage,
+      });
+
+      await message.channel.send({ embeds: [embed] });
     } catch (error) {
-      console.error("Error in messageCreate event (auto-delete):", error);
+      console.error("Error in messageCreate handler:", error);
     }
   },
 };
